New wording for thenable resolution
The v1.1 language for the Promise Resolution Procedure dictates steps for an
implementation rather than the rules that an implementation must follow. The
rest of the spec is in the form of a set of rules, leaving the implementation
details up to the implementer. Whatever sausage is under the hood, as long as
the test suite is past, it is a conformant implementation.

This attempts to normalize the language into the set-of-rules style.